backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
arrow==1.4.0
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.14.3
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
deep-translator==1.11.4
Deprecated==1.3.1
Django==6.0.2
django-appconf==1.2.0
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-redis==6.0.0
django-taggit==6.1.0
django-tailwind==4.4.2
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
frozenlist==1.8.0
honcho==2.0.0
idna==3.11
jaconv==0.5.0
Jinja2==3.1.6
korean_romanizer==0.28.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
mysql-connector-python==9.6.0
mysqlclient==2.2.8
openai==0.27.10
packaging==26.0
propcache==0.4.1
Pygments==2.19.2
pykakasi==2.3.0
pypinyin==0.55.0
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse2==1.7.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.3.1
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
sqlglot==28.10.1
sqlparse==0.5.5
tabulate==0.9.0
text-unidecode==1.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
wrapt==2.1.1
yarl==1.22.0
